Maps both type and kind to C-types in order to fix errors when a Fortran real is declared as `real(rt)` with `rt` imported from a module.
In this case f90wrap would convert `real(rt)` to `real(8)` correctly
but would then identify this as a 32 bit numpy float instead of a 64
bit numpy float. It was not possible to change this with a map file
because `8` was not recognized as a valid string key for the map
dictionary.